@@ -1,31 +1,48 @@
 buildscript {
     repositories {
+        google()
+        jcenter()
         mavenCentral()
     }
 
-    dependencies {
-        classpath 'me.tatarka:gradle-retrolambda:3.7.0'
-    }
+//    dependencies {
+//        classpath 'me.tatarka:gradle-retrolambda:3.7.0'
+//    }
 }
 
 // Required because retrolambda is on maven central
 repositories {
+    google()
+    jcenter()
     mavenCentral()
+//    maven {
+//        url 'https://maven.google.com'
+//    }
+    maven {
+        url 'https://oss.sonatype.org/content/repositories/snapshots/'
+    }
 }
 
 apply plugin: 'com.android.application'
-apply plugin: 'me.tatarka.retrolambda'
+//apply plugin: 'me.tatarka.retrolambda'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId 'org.rares.miner49er'
         minSdkVersion 14
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName '1.0'
-        testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+
+        javaCompileOptions {
+            annotationProcessorOptions {
+//                classNames ["butterknife.compiler.ButterKnifeProcessor"]
+                includeCompileClasspath false
+            }
+        }
     }
     buildTypes {
         release {
@@ -36,22 +53,30 @@ android {
     compileOptions {
         android.defaultConfig.vectorDrawables.useSupportLibrary = true
         sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_7
+        targetCompatibility JavaVersion.VERSION_1_8
     }
+
 }
 // Configuration 'compile' is obsolete and has been replaced with 'implementation' and 'api'.
 dependencies {
     //    retrolambdaConfig 'net.orfjackal.retrolambda:retrolambda:2.5.4'
     implementation fileTree(include: ['*.jar', '*.aar'], dir: 'libs')
     // include ratv_lib.aar here.
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    // include my butter knife 881
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    implementation 'com.jakewharton:butterknife:8.8.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:27.1.1'
+
+//    implementation 'com.jakewharton:butterknife:8.8.1'
+//    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    implementation 'com.jakewharton:butterknife:9.0.0-SNAPSHOT'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:9.0.0-SNAPSHOT'
+//    annotationProcessor 'libs/butterknife-compiler-8.8.1.jar'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    annotationProcessor 'androidx.annotation:annotation:1.0.0'  //  !
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha2'
+    implementation 'com.google.android.material:material:1.1.0-alpha01'
     testImplementation 'junit:junit:4.12'
     compileOnly 'org.projectlombok:lombok:1.16.20'
     annotationProcessor 'org.projectlombok:lombok:1.16.20'
