@@ -1,13 +1,25 @@
-package org.rares.miner49er.domain.issues;
+package org.rares.miner49er.domain.issues.ui.control;
 
-import android.support.v7.widget.RecyclerView;
+import android.view.MenuItem;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.FragmentManager;
+import androidx.recyclerview.widget.RecyclerView;
 import org.rares.miner49er.BaseInterfaces.DomainLink;
+import org.rares.miner49er.R;
 import org.rares.miner49er._abstract.AbstractAdapter;
 import org.rares.miner49er._abstract.ItemViewProperties;
 import org.rares.miner49er._abstract.ResizeableItemViewHolder;
 import org.rares.miner49er._abstract.ResizeableItemsUiOps;
 import org.rares.miner49er.domain.issues.adapter.IssuesAdapter;
+import org.rares.miner49er.domain.issues.repository.IssuesRepository;
 import org.rares.miner49er.layoutmanager.ResizeableLayoutManager;
+import org.rares.miner49er.ui.actionmode.GenericMenuActions;
+import org.rares.miner49er.ui.actionmode.ToolbarActionManager;
+
+import static org.rares.miner49er.ui.actionmode.ToolbarActionManager.MenuConfig.FLAGS;
+import static org.rares.miner49er.ui.actionmode.ToolbarActionManager.MenuConfig.ICON_ID;
+import static org.rares.miner49er.ui.actionmode.ToolbarActionManager.MenuConfig.ITEM_ID;
 
 /**
  * @author rares
@@ -16,15 +28,81 @@ import org.rares.miner49er.layoutmanager.ResizeableLayoutManager;
 
 public class IssuesUiOps extends ResizeableItemsUiOps
         implements
+        ToolbarActionManager.MenuActionListener,
         DomainLink {
 
     private static final String TAG = IssuesUiOps.class.getSimpleName();
     private IssuesRepository issuesRepository = new IssuesRepository();
 
+    private ToolbarActionManager toolbarManager = null;
+
     public IssuesUiOps(RecyclerView rv) {
         setRv(rv);
         issuesRepository.setup();
         repository = issuesRepository;
+    }
+
+    @Override
+    public boolean onListItemClick(ResizeableItemViewHolder holder) {
+        final ListState state = getRvState();
+        boolean enlarge = super.onListItemClick(holder);
+
+        if (toolbarManager == null) {
+            Toolbar t = ((AppCompatActivity) getRv().getContext()).findViewById(R.id.toolbar_c);
+            toolbarManager = (ToolbarActionManager) t.getTag(R.integer.tag_toolbar_action_manager);
+        }
+
+        if (enlarge) {
+            toolbarManager.removeActionListener(this);
+        } else {
+            if (state == ListState.LARGE) {
+                // the toolbar should already be in action mode
+                toolbarManager.addActionListener(this);
+            }
+            toolbarManager.refreshActionMode();
+        }
+        return enlarge;
+    }
+
+    @Override
+    public boolean onToolbarBackPressed() {
+        IssuesAdapter adapter = (IssuesAdapter) getRv().getAdapter();
+        ResizeableItemViewHolder holder =
+                (ResizeableItemViewHolder) getRv().findViewHolderForAdapterPosition(adapter.getLastSelectedPosition());
+        final ListState beforeState = getRvState();
+        onListItemClick(holder);
+        return !beforeState.equals(getRvState());
+    }
+
+    @Override
+    public void configureCustomActionMenu(ToolbarActionManager.MenuConfig config) {
+        // for now, re-use the generic one configured in the projects
+        ResizeableItemViewHolder holder = getSelectedViewHolder();
+        if (holder != null) {
+//            config.title = holder.getLongTitle();
+//            config.subtitle = holder.getInfoLabelString();
+            config.subtitle = holder.getLongTitle();
+            config.subtitleRes = 0;
+        }
+
+        config.createGenericMenu = true;
+
+        config.additionalMenuId = R.menu.menu_additional_issues;
+        config.additionalResources = new int[1][3];
+
+        config.additionalResources[0][ITEM_ID] = R.id.action_set_auto_add_hours;
+        config.additionalResources[0][ICON_ID] = R.drawable.icon_path_auto_add_time_2;
+        config.additionalResources[0][FLAGS] = MenuItem.SHOW_AS_ACTION_NEVER;
+    }
+
+    @Override
+    public GenericMenuActions getMenuActionsProvider() {
+        return null;
+    }
+
+    @Override
+    protected void configureMenuActionsProvider(FragmentManager fm) {
+
     }
 
     @Override
