@@ -1,6 +1,6 @@
 package org.rares.miner49er.persistence.resolvers;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import com.pushtorefresh.storio3.sqlite.operations.delete.DefaultDeleteResolver;
 import com.pushtorefresh.storio3.sqlite.queries.DeleteQuery;
 import org.rares.miner49er.persistence.entities.TimeEntry;
