@@ -1,7 +1,7 @@
 package org.rares.miner49er.persistence.resolvers;
 
 import android.database.Cursor;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import com.pushtorefresh.storio3.sqlite.StorIOSQLite;
 import com.pushtorefresh.storio3.sqlite.operations.get.DefaultGetResolver;
 import org.rares.miner49er.persistence.entities.TimeEntry;
