@@ -1,9 +1,9 @@
 package org.rares.miner49er._abstract;
 
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import io.reactivex.functions.Consumer;
 import lombok.Getter;
 import lombok.Setter;
