@@ -2,7 +2,7 @@ package org.rares.miner49er._abstract;
 
 import lombok.Data;
 import org.rares.miner49er.domain.entries.adapter.TimeEntryViewProperties;
-import org.rares.miner49er.domain.issues.adapter.IssuesViewProperties;
+import org.rares.miner49er.domain.issues.ui.viewholder.IssuesViewProperties;
 import org.rares.miner49er.domain.projects.adapter.ProjectViewProperties;
 import org.rares.miner49er.persistence.entities.Issue;
 import org.rares.miner49er.persistence.entities.TimeEntry;
