@@ -1,10 +1,10 @@
 package org.rares.miner49er.layoutmanager.postprocessing.resize;
 
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import androidx.recyclerview.widget.RecyclerView;
 import org.rares.miner49er._abstract.AbstractAdapter;
 import org.rares.miner49er.layoutmanager.postprocessing.ResizePostProcessor;
 import org.rares.ratv.rotationaware.RotationAwareTextView;
