@@ -1,9 +1,9 @@
 package org.rares.miner49er.layoutmanager.postprocessing.rotation;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import androidx.recyclerview.widget.RecyclerView;
 import org.rares.miner49er.R;
 import org.rares.miner49er._abstract.AbstractAdapter;
 
