@@ -1,11 +1,11 @@
 package org.rares.miner49er.util;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.Gravity;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+import androidx.recyclerview.widget.RecyclerView;
 import org.rares.miner49er._abstract.ResizeableItemViewHolder;
 import org.rares.ratv.rotationaware.RotationAwareTextView;
 
